backend/requirements.txt
flask==3.0.3
supabase==2.9.0
python-dotenv==1.0.1
flask-cors==4.0.1
pdfplumber==0.11.4
spacy==3.7.6
whoosh==2.7.4
gunicorn==23.0.0  # For deployment